Added the auth service
The service is able to register and login user with email or google Auth,
then to redirect it based on the dynamic parameter, or in case of register - to app view.
Depending on user env (desktop or mobile) it uses popup login or redirect (google Auth).
After Register/login it checks if the user exists in db and returns the valid action based on it.